Update 11-software-project.md
Hi There - I would like to make a suggestion :) I am a novice, and on Linux, so what I am suggesting may not be what you had in mind. 
Line 68: If I am reading this correctly, the lesson assumes that my Home folder has already been "git init" enabled. So that I would be able to use the "git clone" command. However if I "git init" my whole Home directory, would I not have my entire Home folder listed in a "git status" ? My suggestion is for students to make a directory, cd into that directory, and then "git init" just that directory, before using the clone command.  Also I noticed that the command "git clone git@github.com:<YOUR_GITHUB_USERNAME>/python-intermediate-inflammation" does not have a .git on the end of it ? When I copy it from GitHub , it has a .git at the end. I don't know if that is important or not. Many Thanks. I look forward to the rest of the lesson :)