Don't use exception handling for control flow in OOP example
It is an anti-pattern to use exceptions to manage control flow. In this
specific case it would be bad as an observation without a 'day' would
silently just make the next observation start at day 0. In general it is
a bad practise as is essentially a goto so can make code harder to
follow and can make debugging harder (if auto breaking on exceptions).