Make array indentation in examples consistent with GH Actions scripts
In the GitHub Actions the elements of arrays are not indented relative
to the key of the array. Both formats are acceptable in YAML but to keep
this tutorial as simple as possible stick to one.

Opted for the no indentation as this is what GitHub will generate for
you if you create a new workflow.