Remove advice about force pushing
While this is generally good advice, some open source repositories will
require force-pushing to keep commit history tidy. Further, this is good
advice in general, not related to pull requests, and most people won't
have come across the idea of force-pushing anway, so no need to confuse
them.

It also isn't accurate, it can't corrupt the pull request, it can just
get the reviewer tied in a bit of a knot if they don't know what they
are doing and pulled the code locally.