Incorporate the initial exercise more fully into the final exercise
Add in a quesiton about the benifits of code review to ensure the
benifits have been effectively communicated by this module.

Include the benfitis from the solution to the initial exercise to ensure
the key points are reinforced at the end of the module.

Also include some example problems to stimulate discussion on the
challenges of having a code review process.