Remove section 3 slide: 'Merging the Feature In'
The first task in section 4 is a code review exercise, which instructs
the students to open a pull request on GitHub. However, this slide
at the end of section 3 instructs the students to merge their changes
into develop locally.

This caused some confusion last time I ran this, as the students had no
existing changes to make a pull request from!